// src/app/application/lead-management/models/lead-status.enum.ts
export enum LeadStatus {
  NEW = 'NEW',
  CONTACTED = 'CONTACTED',
  INTERESTED = 'INTERESTED',
  FOLLOW_UP = 'FOLLOW_UP',
  ADMISSION_CONFIRMED = 'ADMISSION_CONFIRMED',
  LOST = 'LOST'
}

export enum LeadSource {
  WEBSITE = 'WEBSITE',
  SOCIAL_MEDIA = 'SOCIAL_MEDIA',
  WALK_IN = 'WALK_IN',
  PHONE_CALL = 'PHONE_CALL',
  REFERRAL = 'REFERRAL'
}

export enum InteractionType {
  CALL = 'CALL',
  WHATSAPP = 'WHATSAPP',
  EMAIL = 'EMAIL',
  SMS = 'SMS',
  PERSONAL_MEETING = 'PERSONAL_MEETING'
}

export enum ContactOutcome {
  CONNECTED = 'CONNECTED',
  NOT_CONNECTED = 'NOT_CONNECTED',
  NO_ANSWER = 'NO_ANSWER',
  WRONG_NUMBER = 'WRONG_NUMBER',
  BUSY = 'BUSY'
}

export enum LostLeadReason {
  FEE_TOO_HIGH = 'FEE_TOO_HIGH',
  JOINED_ANOTHER_INSTITUTE = 'JOINED_ANOTHER_INSTITUTE',
  NOT_ELIGIBLE = 'NOT_ELIGIBLE',
  NO_RESPONSE = 'NO_RESPONSE',
  JOINED_OUR_INSTITUTE = 'JOINED_OUR_INSTITUTE',
  NOT_INTERESTED = 'NOT_INTERESTED'
}